requirements.txt
pytest==7.3.1
requests==2.31.0
allure-pytest==2.13.2
jsonschema==4.17.3
urllib3==2.0.2
deepdiff==6.3.0
python-dotenv==1.0.0